+++ .jbang/JabSrvLauncher.java
@@ -17,7 +17,7 @@
 
 // from jabsrv
 //DEPS com.fasterxml.jackson.core:jackson-annotations:2.21
-//DEPS com.github.ben-manes.caffeine:caffeine:3.2.3
+//DEPS com.github.ben-manes.caffeine:caffeine:3.2.4
 //DEPS com.google.guava:guava:33.6.0-jre
 //DEPS com.konghq:unirest-modules-gson:4.8.1
 //DEPS de.undercouch:citeproc-java:3.5.0